producers/python/requirements.txt
kafka-python==2.0.2
python-dateutil==2.9.0.post0